requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
plotly==5.14.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
Werkzeug==2.3.0
zipp==3.15.0
psycopg2-binary
python-dotenv==1.0.0